# openapi-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SmartUnderwrite.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "paths": {
    "/api/Affiliates": {
      "get": {
        "tags": [
          "Affiliates"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AffiliateDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AffiliateDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AffiliateDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Affiliates"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAffiliateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAffiliateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAffiliateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AffiliateDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AffiliateDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AffiliateDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Affiliates/{id}": {
      "get": {
        "tags": [
          "Affiliates"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AffiliateDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AffiliateDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AffiliateDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Affiliates"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAffiliateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAffiliateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAffiliateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AffiliateDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AffiliateDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AffiliateDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Affiliates/{id}/deactivate": {
      "post": {
        "tags": [
          "Affiliates"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AffiliateDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AffiliateDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AffiliateDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Affiliates/{id}/activate": {
      "post": {
        "tags": [
          "Affiliates"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AffiliateDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AffiliateDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AffiliateDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Affiliates/{id}/users": {
      "get": {
        "tags": [
          "Affiliates"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AffiliateUserDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AffiliateUserDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AffiliateUserDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Affiliates"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignUserToAffiliateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignUserToAffiliateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AssignUserToAffiliateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AffiliateUserDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AffiliateUserDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AffiliateUserDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Affiliates/{id}/users/{userId}": {
      "delete": {
        "tags": [
          "Affiliates"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AffiliateDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "externalId": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "userCount": {
            "type": "integer",
            "format": "int32"
          },
          "applicationCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "AffiliateUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
      "AssignUserToAffiliateRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "CreateAffiliateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "externalId": {
            "type": "string"
          }
        }
      },
...
      "UpdateAffiliateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "externalId": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
...
}